add: information header and file rename

+++ Setup.lua
@@ -1,9 +1,23 @@
--- Pad numbers with leading zeros.
+-- Advent of Code Utility Script
+-- Author: José Areia
+-- Created: 2024-12-01
+-- Description: Automates content generation for a specific Advent of Code day.
+-- Note: Input files are not auto-downloaded intentionally, as I enjoy examining the input manually.
+
 local function pad_number(num)
     return string.format("%02d", num)
 end
 
--- Simulated progress bar
+local function directory_exists(path)
+    local file = io.open(path, "r")
+    if file then
+        file:close()
+        return true
+    else
+        return false
+    end
+end
+
 local function show_progress(verb, task)
     for i = 1, 100 do
         io.write(string.format("\r[\27[32m OK \27[0m] \27[90m%s\27[0m %s... %d%%", verb, task, i))
@@ -13,10 +27,15 @@ local function show_progress(verb, task)
     io.write(string.format("\r[\27[32m OK \27[0m] \27[90m%s\27[0m %s... Done\n", verb, task))
 end
 
--- Create the necessary files inside the directory.
 local function create_perl_files(number)
     local folder_name = pad_number(number)
     show_progress("Creating", "the project directory")
+
+    if not directory_exists(folder_name) then
+        print("Failed to create the directory '" .. folder_name .. "'.")
+        return
+    end
+
     os.execute("mkdir -p " .. folder_name)
 
     local input_file = io.open(folder_name .. "/input.txt", "w")
@@ -40,21 +59,17 @@ local function create_perl_files(number)
 use strict;
 use warnings;
 
-my $input = "input.txt";
-
-open(my $fh, '<', $input) or die "Cannot open file '$input': $!";
+open(my $in, '<', 'input.txt') or die $!;
 
 my @line;
-my $var = 0;
+my $total = 0;
 
-while (<$fh>) {
+while (<$in>) {
     chomp;
     my @line = split(/\s+/, $_);
 }
 
-close($fh);
-
-print "Part ]] .. pad_number(i) .. [[ : $var\n";
+print "Part ]] .. pad_number(i) .. [[ : $total\n";
 ]]
         pl_file:write(script_content)
         pl_file:close()
@@ -68,7 +83,7 @@ end
 io.write("Advent of Code - Please specify the day (1-25): ")
 local number = tonumber(io.read())
 
-if number then
+if number and number >= 1 and number <= 25 then
     create_perl_files(number)
 else
     print("Invalid number entered.")
